fix: error when query by level of s2 geometry

+++ server/src/main/java/me/nearby/adapter/rest/userlocation/dto/NearbyQuery.java
@@ -14,5 +14,5 @@ import java.io.Serializable;
 public class NearbyQuery implements Serializable {
     Double latitude;
     Double longitude;
-    Double radius;
+    Integer radius;
 }
